+++ src/tags/information-for-account-owner.ts
@@ -1,5 +1,5 @@
 import {compareArrays} from './../utils';
-import {colonSymbolCode} from './../tokens';
+import {colonSymbolCode, returnSymbolCode, newLineSymbolCode, spaceSymbolCode} from './../tokens';
 import {Tag, State, Statement} from './../typings';
 
 /**
@@ -21,10 +21,28 @@ const informationTag: Tag = {
 
     close (state: State) {
         const statement: Statement = state.statements[state.statementIndex];
+        const description: number[] = [];
+        let descriptionLength: number = 0;
+
+        // filter denied symbols
+        for (let i = state.tagContentStart; i < state.tagContentEnd; i++) {
+            const symbolCode: number = state.data[i];
+
+            if (
+                // remove \r & \n
+                symbolCode !== returnSymbolCode && symbolCode !== newLineSymbolCode && (
+                    // use 1 space instead of multiple ones
+                    symbolCode !== spaceSymbolCode || description[descriptionLength - 1] !== symbolCode
+                )
+            ) {
+                description[descriptionLength] = symbolCode;
+                descriptionLength++;
+            }
+        }
 
         statement.transactions[state.transactionIndex].description = String.fromCharCode.apply(
             String,
-            state.data.slice(state.tagContentStart, state.tagContentEnd)
+            description
         ).trim();
     }
 };