+++ CHANGELOG.md
@@ -4,3 +4,8 @@
 ## 0.2.0
 * added `npm-shrinkwrap.json`
 * added `Transaction`, `BalanceInfo` interfaces to `.d.ts` file
+
+## 0.3.0
+* Transaction description
+    * filter symbols `\r`, `\n`
+    * replace multiple spaces by 1 space symbol